requirements.txt
fastapi>=0.95.0,<0.110.0
uvicorn>=0.22.0,<0.28.0
python-multipart>=0.0.6
aiohttp>=3.8.4
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.3.0,<1.4.0
aiofiles>=23.1.0
httpx>=0.24.0,<0.25.0
websockets>=11.0.0
ffmpeg-python>=0.2.0
PyYAML>=6.0 